+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Version 1.2.2 (not released yet)
 - fix invalid player entity model when undercover (RF bug)
 - speed up server list refreshing
 - add linear_pitch command for making pitch control linear
+- allow 'level' command outside of multiplayer game and remove now redundant 'levelsp' command
 - fix multiple security vulnerabilities
 - stability improvements